requirements.txt
hydra-core==1.1.0
PyAudio==0.2.11
pytest==6.2.4